Refactored into survey folder
