+++ client/src/components/survey/question.js
@@ -6,7 +6,6 @@ const Question = ({ question, length }) => {
     <div>
       <h1>QUESTION</h1>
       {jaba}
-      // Wathc out for merge conflicts >:DD
       {question}
       {length}
     </div>
